// src/api/models/education-im.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface EducationIM
 */
export interface EducationIM {
    /**
     * 
     * @type {string}
     * @memberof EducationIM
     */
    resumeId: string;
    /**
     * 
     * @type {string}
     * @memberof EducationIM
     */
    schoolName: string;
    /**
     * 
     * @type {string}
     * @memberof EducationIM
     */
    degree: string;
    /**
     * 
     * @type {string}
     * @memberof EducationIM
     */
    fieldOfStudy: string;
    /**
     * 
     * @type {string}
     * @memberof EducationIM
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof EducationIM
     */
    startDate: string;
    /**
     * 
     * @type {string}
     * @memberof EducationIM
     */
    endDate?: string | null;
}