// src/api/models/education-um.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface EducationUM
 */
export interface EducationUM {
    /**
     * 
     * @type {string}
     * @memberof EducationUM
     */
    schoolName: string;
    /**
     * 
     * @type {string}
     * @memberof EducationUM
     */
    degree: string;
    /**
     * 
     * @type {string}
     * @memberof EducationUM
     */
    fieldOfStudy: string;
    /**
     * 
     * @type {string}
     * @memberof EducationUM
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof EducationUM
     */
    startDate: string;
    /**
     * 
     * @type {string}
     * @memberof EducationUM
     */
    endDate?: string | null;
}